WIP: internal state lock protects other "locks"
etcd mutexes are (rightly) lost if the client holding them
disconnects. Thus, since we want to have mode that hold locks
longer than our client runs, we need a way to mark things locked
in the persistent key-value part of etcd. This is the first
step in this direction. A well-known "internal state lock" is
now used to protect all changes of persistent etcd-store
distlock data.

This is a work in progress. It doesn't break anything, but
it doesn't give us any new functionality, either.