Chnaged internal storage format, prettier lists
The internal way of storing lock information was changed from
using two keys (reason and timestamp) to only one key and a
compound value, which is cleaner and means we will not neet
transactions to store and delete cleanly.

Also, the output format of --list has been improved again, making
it more human-readable. This was done in lock-step with switching
the internal storage format, so it's in the same commit.