New environment variable DISTLOCK_PREFIX
The internal namespace distlock uses inside of etcd to store it's
data starts with a prefix that used to be harcoded to "/distlock/".
It can now be changed using the environment variable
DISTLOCK_PREFIX. This might be useful if you want to run several
distlock environments against the same etcd, like for example a
test, a staging and a production environment.