Added config file functionality
Settings that can be set via command line or environment variables
can now also be set in a config file. The file must be in YAML
format. The default path is "/etc/distlock/distlock.yaml" but can
be changed using the environment variable DISTLOCK_CONFIG.

The config file also allows setting the interal lock name, but
does not allow specifying a command to execute.